Moving database classes into seperate modules
