+++ src/tests/test_get_dataset.py
@@ -4,7 +4,7 @@ from sqlalchemy import Engine
 from sqlalchemy.orm import Session
 from starlette.testclient import TestClient
 
-from database.models import OrmDataset
+from database.model.dataset import OrmDataset
 
 OPENML_URL = "https://www.openml.org/api/v1/json"
