+++ src/tests/test_get_datasets.py
@@ -2,7 +2,7 @@ from sqlalchemy import Engine
 from sqlalchemy.orm import Session
 from starlette.testclient import TestClient
 
-from database.models import OrmDataset
+from database.model.dataset import OrmDataset
 
 
 def test_happy_path(client: TestClient, engine: Engine):
