+++ src/tests/test_get_publications.py
@@ -2,7 +2,7 @@ from sqlalchemy import Engine
 from sqlalchemy.orm import Session
 from starlette.testclient import TestClient
 
-from database.models import OrmPublication
+from database.model.publication import OrmPublication
 
 
 def test_happy_path(client: TestClient, engine: Engine):
